third_party/WebKit/LayoutTests/external/wpt/html/webappapis/scripting/processing-model-2/compile-error-in-setTimeout-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected end of input
This is a testharness.js-based test.
FAIL window.onerror - compile error in setTimeout assert_equals: second arg expected "http://web-platform.test:8001/html/webappapis/scripting/processing-model-2/compile-error-in-setTimeout.html" but got ""
PASS window.onerror - compile error in setTimeout (column) 
Harness: the test ran to completion.
